REFACTOR LOGIN PAGE NEW FORMAT FOR JAVA
A refactor of folders and classes was performed. Non-HTML elements were
moved to a static folder because HTML in Java is dynamic and constantly
changes. If static elements like images are included directly, it makes
rendering the HTML more difficult. Annotations were added to make the
HTML work with Java, and the design file was modified as per the
request. The button now uses a submit action instead of being a
reference.